adversarial_detector/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.19.0
PyYAML>=5.4.1
streamlit>=1.22.0
matplotlib>=3.5.0
seaborn>=0.11.0
